added functions

+++ src/main/java/com/example/onlineschoolapp/dto/StudentDTO.java
@@ -12,5 +12,5 @@ public class StudentDTO {
     private String firstName;
     private String lastName;
     private String email;
-    private Integer age;
+    private Double age;
 }
